+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 All notable changes to the "robotcode" extension will be documented in this file.
 
-## [Unreleased]
-- none so far
+## [Unreleased]
+
+- Reimplement workspace analysis
+- Optimize the search for unused references
+
 ##  0.13.19
 
 - Add a the setting `robotcode.completion.filterDefaultLanguage` to filter english language in completion, if there is another language defined for workspace or in file